wayland: refactor break out XWindow,XStatic to xwindow.py
